Improve drag event handling
Fine adjustment of vertices is prevented by is_click when they are
within 4px of the starting position. This is to prevent imprecise
clicks from being interpretted as drags, but also prevents fine
adjustment. This commit adds a timeout to these filters so that fine adjustments
can be made after 1 second of pause.